Move r_misc into its own js module
