+++ src/smc-webapp/r_misc/__tests__/close-x.test.tsx
