+++ src/smc-webapp/r_misc/__tests__/icon.test.tsx
