+++ src/smc-webapp/r_misc/close-x.tsx
