+++ src/smc-webapp/r_misc/index.cjsx
@@ -20,12 +20,12 @@
 
 async = require('async')
 
-{Component, React, ReactDOM, rclass, rtypes, is_redux, is_redux_actions, redux, Store, Actions, Redux} = require('./app-framework')
+{Component, React, ReactDOM, rclass, rtypes, is_redux, is_redux_actions, redux, Store, Actions, Redux} = require('../app-framework')
 {Alert, Button, ButtonToolbar, Checkbox, Col, FormControl, FormGroup, ControlLabel, InputGroup, Overlay, OverlayTrigger, Popover, Modal, Tooltip, Row, Well} = require('react-bootstrap')
-{HelpEmailLink, SiteName, CompanyName, PricingUrl, PolicyTOSPageUrl, PolicyIndexPageUrl, PolicyPricingPageUrl} = require('./customize')
-{UpgradeRestartWarning} = require('./upgrade_restart_warning')
+{HelpEmailLink, SiteName, CompanyName, PricingUrl, PolicyTOSPageUrl, PolicyIndexPageUrl, PolicyPricingPageUrl} = require('../customize')
+{UpgradeRestartWarning} = require('../upgrade_restart_warning')
 copy_to_clipboard = require('copy-to-clipboard')
-{reportException} = require('../webapp-lib/webapp-error-reporter')
+{reportException} = require('../../webapp-lib/webapp-error-reporter')
 
 {Icon} = require('./icon')
 exports.Icon = Icon
@@ -52,12 +52,12 @@ theme       = require('smc-util/theme')
 immutable   = require('immutable')
 underscore  = require('underscore')
 
-markdown    = require('./markdown')
-feature     = require('./feature')
+markdown    = require('../markdown')
+feature     = require('../feature')
 
 {defaults, required} = misc
 
-exports.MarkdownInput = require('./widget-markdown-input/main').MarkdownInput
+exports.MarkdownInput = require('../widget-markdown-input/main').MarkdownInput
 
 # base unit in pixel for margin/size/padding
 exports.UNIT = UNIT = 15
@@ -812,9 +812,9 @@ exports.HTML = HTML = rclass
             return {__html: ''}
 
         if @props.safeHTML
-            html = require('./misc_page').sanitize_html_safe(@props.value, @props.post_hook)
+            html = require('../misc_page').sanitize_html_safe(@props.value, @props.post_hook)
         else
-            html = require('./misc_page').sanitize_html(@props.value, true, true, @props.post_hook)
+            html = require('../misc_page').sanitize_html(@props.value, true, true, @props.post_hook)
 
         if exports.SHARE_SERVER
             {jQuery} = require('smc-webapp/jquery-plugins/katex')  # ensure have plugin here.
@@ -1133,7 +1133,7 @@ exports.DeletedProjectWarning = ->
 exports.course_warning = (pay) ->
     if not pay
         return false
-    {webapp_client} = require('./webapp_client')
+    {webapp_client} = require('../webapp_client')
     return webapp_client.server_time() <= misc.months_before(-3, pay)  # require subscription until 3 months after start (an estimate for when class ended, and less than when what student did pay for will have expired).
 
 project_warning_opts = (opts) ->
@@ -1162,13 +1162,13 @@ exports.CourseProjectWarning = (opts) ->
         return <span></span>
     # We may now assume course_info.get is defined, since course_warning is only true if it is.
     pay = course_info.get('pay')
-    billing = require('./billing')
+    billing = require('../billing')
     if avail > 0
         action = <billing.BillingPageLink text="move this project to a members only server" />
     else
         action = <billing.BillingPageLink text="buy a course subscription" />
     is_student = account_id == course_info.get('account_id') or email_address == course_info.get('email_address')
-    {webapp_client} = require('./webapp_client')
+    {webapp_client} = require('../webapp_client')
     if pay > webapp_client.server_time()  # in the future
         if is_student
             deadline  = <span>Your instructor requires you to {action} within <TimeAgo date={pay}/>.</span>
@@ -1329,7 +1329,7 @@ exports.EditorFileInfoDropdown = EditorFileInfoDropdown = rclass
                 'copy'     : 'files-o'
         else
             # dynamically create a map from 'key' to 'icon'
-            {file_actions} = require('./project_store')
+            {file_actions} = require('../project_store')
             items = underscore.object(([k, v.icon] for k, v of file_actions))
 
         for name, icon of items
@@ -1374,7 +1374,7 @@ exports.NoUpgrades = NoUpgrades = rclass
 
     billing: (e) ->
         e.preventDefault()
-        require('./billing').visit_billing_page()
+        require('../billing').visit_billing_page()
 
     render: ->
         <Alert bsStyle='info'>
