+++ src/smc-webapp/r_misc/space.tsx
